aihr_test/backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
dashscope
pydantic
